+++ internal/provider/csm/metadata.go
@@ -17,40 +17,9 @@ type NodeMetadata struct {
 	AdditionalProperties map[string]interface{}
 }
 
-// func (nm *NodeMetadata) Validate() error {
-// 	var validationErrs []error
-// 	// Role checks
-// 	if nm.Role == nil {
-// 		validationErrs = append(validationErrs, fmt.Errorf("no role set"))
-// 	}
-
-// 	// Check to see if the role matches what is allowed by HSM
-// 	// TODO
-
-// 	// Check to see if the sub-role matches what is allowed by HSM
-
-// 	// NID checks
-// 	if nm.Nid == nil {
-// 		validationErrs = append(validationErrs, fmt.Errorf("no nid set"))
-// 	}
-// 	if *nm.Nid < 0 {
-// 		validationErrs = append(validationErrs, fmt.Errorf("a non-positive NID is set: %d", *nm.Nid))
-// 	}
-
-// 	// Alias checks
-// 	if nm.Alias == nil {
-// 		validationErrs = append(validationErrs, fmt.Errorf("no alias set"))
-// 	}
-// 	if strings.Contains(*nm.Alias, " ") {
-// 		validationErrs = append(validationErrs, fmt.Errorf("alias contains spaces"))
-// 	}
-
-// 	if len(validationErrs) == 0 {
-// 		return nil
-// 	}
-
-// 	return errors.Join(validationErrs...)
-// }
+type CabinetMetadata struct {
+	HMNVlan *int
+}
 
 // TODO this might need a better home
 func StringPtr(s string) *string {
